src/requirements.txt
asgiref==3.8.1
boto3==1.34.131
botocore==1.34.131
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.6
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-jazzmin==3.0.0
django-modeltranslation==0.19.4
django-seed==0.3.1
django-storages==1.14.3
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
Faker==28.0.0
gunicorn==22.0.0
idna==3.7
jmespath==1.0.1
mutagen==1.47.0
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.1
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
toposort==1.10
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2